chore: usecase interface로 분리
