+++ src/main/java/com/example/demo/PageService.java
@@ -1,0 +1,38 @@
+package com.example.demo;
+
+import com.example.demo.persistence.LoadPageInfoCommand;
+import com.example.demo.persistence.Page;
+import com.example.demo.persistence.SummaryPageInfo;
+
+import java.util.ArrayList;
+import java.util.List;
+
+class PageService implements PageUseCase {
+    private final LoadPageInfoCommand loadPageInfoCommand;
+
+    public PageService(final LoadPageInfoCommand loadPageInfoCommand) {
+        this.loadPageInfoCommand = loadPageInfoCommand;
+    }
+
+    @Override
+    public PageResponse getPageView(final PageViewRequest request) {
+        final Page targetPage = loadPageInfoCommand.selectAllAttributes(request.pageId());
+        final List<SummaryPageInfo> subPages = loadSubPages(request.pageId());
+        final List<SummaryPageInfo> breadcrumbs = loadBreadcrumbs(targetPage.parentId());
+        return PageResponse.of(targetPage, subPages, breadcrumbs);
+    }
+
+    protected List<SummaryPageInfo> loadBreadcrumbs(final String pageId) {
+        SummaryPageInfo summaryOfParentPage = loadPageInfoCommand.selectSummaryOfParentPage(pageId);
+        if (summaryOfParentPage.parentId() == null) {
+            return new ArrayList<>(List.of(summaryOfParentPage));
+        }
+        final List<SummaryPageInfo> summaryPageInfoList = loadBreadcrumbs(summaryOfParentPage.parentId());
+        summaryPageInfoList.add(summaryOfParentPage);
+        return summaryPageInfoList;
+    }
+
+    private List<SummaryPageInfo> loadSubPages(final String pageId) {
+        return loadPageInfoCommand.selectSummaryOfSubPages(pageId);
+    }
+}
